feat: improve sidebar styling, add dropdown, padding and margins
