+++ src/app/(website)/docs/Sidebar.module.css
@@ -1,8 +1,28 @@
-.menu {
-  font-size: 14px;
+.sidebar {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: calc(100vh - 150px);
+  top: 121px;
   position: sticky;
-  top: 70px;
-  margin-top: 70px;
+}
+
+.menuWrapper {
+  position: relative;
+  overflow: hidden;
+}
+
+.dropdown {
+  margin-bottom: 2rem;
+}
+
+.menu {
+  scrollbar-gutter: stable;
+  overflow-y: scroll;
+  display: flex;
+  flex-direction: column;
+  height: calc(100vh - 200px);
+  font-size: 14px;
   color: var(--base800);
 }
 
@@ -43,18 +63,4 @@
 
 .menu .item.selected a {
   color: var(--primary400);
-}
-
-.menu-wrapper {
-  position: relative;
-  overflow: hidden;
-  padding-right: 1rem;
-}
-
-.menu {
-  scrollbar-gutter: stable;
-  overflow-y: scroll;
-  display: flex;
-  flex-direction: column;
-  height: calc(100vh - 200px);
 }
