+++ src/app/(website)/docs/layout.module.css
@@ -1,6 +1,7 @@
 .docs {
   display: grid;
   grid-template-columns: 250px 1fr;
+  column-gap: 4rem;
   color: var(--base800);
 }
